[ADF-3540] Moved user services to Core module
