+++ lib/core/userinfo/public-api.ts
@@ -22,5 +22,6 @@ export * from './services/identity-user.service';
 export * from './models/bpm-user.model';
 export * from './models/ecm-user.model';
 export * from './models/identity-user.model';
+export * from './models/identity-role.model';
 
 export * from './userinfo.module';
