+++ lib/core/userinfo/services/identity-user.service.ts
@@ -16,9 +16,14 @@
  */
 
 import { Injectable } from '@angular/core';
-import { of, Observable } from 'rxjs';
+import { of, Observable, from } from 'rxjs';
+import { map } from 'rxjs/operators';
+
 import { IdentityUserModel } from '../models/identity-user.model';
-import { JwtHelperService } from './../../services/jwt-helper.service';
+import { JwtHelperService } from '../../services/jwt-helper.service';
+import { AppConfigService } from '../../app-config/app-config.service';
+import { AlfrescoApiService } from '../../services';
+import { IdentityRoleModel } from '../models/identity-role.model';
 
 @Injectable({
     providedIn: 'root'
@@ -30,7 +35,10 @@ export class IdentityUserService {
     static USER_ACCESS_TOKEN = 'access_token';
     static USER_PREFERRED_USERNAME = 'preferred_username';
 
-    constructor(private helper: JwtHelperService) {}
+    constructor(
+        private helper: JwtHelperService,
+        private apiService: AlfrescoApiService,
+        private appConfigService: AppConfigService) {}
 
     getCurrentUserInfo(): Observable<IdentityUserModel> {
         const fullName = this.getValueFromToken<string>(IdentityUserService.USER_NAME);
@@ -50,4 +58,93 @@ export class IdentityUserService {
         }
         return <T> value;
     }
+
+    getUsers(): Observable<IdentityUserModel[]> {
+        const url = this.buildUserUrl();
+        const httpMethod = 'GET', pathParams = {}, queryParams = {}, bodyParam = {}, headerParams = {},
+            formParams = {}, authNames = [], contentTypes = ['application/json'], accepts = ['application/json'];
+
+        return from(this.apiService.getInstance().oauth2Auth.callCustomApi(
+                    url, httpMethod, pathParams, queryParams,
+                    headerParams, formParams, bodyParam, authNames,
+                    contentTypes, accepts, Object, null, null)
+                ).pipe(
+                    map((response: IdentityUserModel[]) => {
+                        return response;
+                    })
+            );
+    }
+
+    getUserRoles(userId: string): Observable<IdentityRoleModel[]> {
+        const url = this.buildRolesUrl(userId);
+        const httpMethod = 'GET', pathParams = {}, queryParams = {}, bodyParam = {}, headerParams = {},
+            formParams = {}, authNames = [], contentTypes = ['application/json'], accepts = ['application/json'];
+
+        return from(this.apiService.getInstance().oauth2Auth.callCustomApi(
+                    url, httpMethod, pathParams, queryParams,
+                    headerParams, formParams, bodyParam, authNames,
+                    contentTypes, accepts, Object, null, null)
+                ).pipe(
+                    map((response: IdentityRoleModel[]) => {
+                        return response;
+                    })
+            );
+    }
+
+    async getUsersByRolesWithCurrentUser(roleNames: string[]): Promise<IdentityUserModel[]> {
+        const filteredUsers: IdentityUserModel[] = [];
+        if (roleNames && roleNames.length > 0) {
+            const users = await this.getUsers().toPromise();
+
+            for (let i = 0; i < users.length; i++) {
+                const hasAnyRole = await this.userHasAnyRole(users[i].id, roleNames);
+                if (hasAnyRole) {
+                    filteredUsers.push(users[i]);
+                }
+            }
+        }
+
+        return filteredUsers;
+    }
+
+    async getUsersByRolesWithoutCurrentUser(roleNames: string[]): Promise<IdentityUserModel[]> {
+        const filteredUsers: IdentityUserModel[] = [];
+        if (roleNames && roleNames.length > 0) {
+            const currentUser = await this.getCurrentUserInfo().toPromise();
+            let users = await this.getUsers().toPromise();
+
+            users = users.filter((user) => { return user.username !== currentUser.username; });
+
+            for (let i = 0; i < users.length; i++) {
+                const hasAnyRole = await this.userHasAnyRole(users[i].id, roleNames);
+                if (hasAnyRole) {
+                    filteredUsers.push(users[i]);
+                }
+            }
+        }
+
+        return filteredUsers;
+    }
+
+    private async userHasAnyRole(userId: string, roleNames: string[]): Promise<boolean> {
+        const userRoles = await this.getUserRoles(userId).toPromise();
+        const hasAnyRole = roleNames.some((roleName) => {
+            const filteredRoles = userRoles.filter((userRole) => {
+                return userRole.name.toLocaleLowerCase() === roleName.toLocaleLowerCase();
+            });
+
+            return filteredRoles.length > 0;
+        });
+
+        return hasAnyRole;
+    }
+
+    private buildUserUrl(): any {
+        return `${this.appConfigService.get('bpmHost')}/auth/admin/realms/springboot/users`;
+    }
+
+    private buildRolesUrl(userId: string): any {
+        return `${this.appConfigService.get('bpmHost')}/auth/admin/realms/springboot/users/${userId}/role-mappings/realm/composite`;
+    }
+
 }
