+++ lib/process-services-cloud/src/lib/start-task-cloud/components/start-task-cloud.component.spec.ts
@@ -60,8 +60,8 @@ describe('StartTaskCloudComponent', () => {
         service = TestBed.get(StartTaskCloudService);
         identityService = TestBed.get(IdentityUserService);
         createNewTaskSpy = spyOn(service, 'createNewTask').and.returnValue(of(taskDetailsMock));
-        getRolesByUserIdSpy = spyOn(service, 'getUserRoles').and.returnValue(of(mockRoles));
-        getUserSpy = spyOn(service, 'getUsers').and.returnValue(of(mockUsers));
+        getRolesByUserIdSpy = spyOn(identityService, 'getUserRoles').and.returnValue(of(mockRoles));
+        getUserSpy = spyOn(identityService, 'getUsers').and.returnValue(of(mockUsers));
         spyOn(identityService, 'getCurrentUserInfo').and.returnValue(of(new IdentityUserModel({username: 'currentUser'})));
         fixture.detectChanges();
     }));
