+++ lib/process-services-cloud/src/lib/start-task-cloud/mock/user-cloud.mock.ts
@@ -3,11 +3,3 @@ export let fakeUsers = [
     { username: 'mock-username2', firstName: 'mock-firstName2', lastName: 'mock-lastName2' },
     { username: 'mock-username3', firstName: 'mock-firstName3', lastName: 'mock-lastName3' }
 ];
-
-export let mockRoles = [
-    { id: 'id-1', name: 'MOCK-ADMIN-ROLE'},
-    { id: 'id-2', name: 'MOCK-USER-ROLE'},
-    { id: 'id-3', name: 'MOCK_MODELER-ROLE' },
-    { id: 'id-4', name: 'MOCK-ROLE-1' },
-    { id: 'id-5', name: 'MOCK-ROLE-2'}
-];
