+++ lib/process-services-cloud/src/lib/start-task-cloud/services/start-task-cloud.service.ts
@@ -15,14 +15,11 @@
  * limitations under the License.
  */
 
-import { RoleCloudModel } from './../models/role-cloud.model';
 import { Injectable } from '@angular/core';
 import {
     AlfrescoApiService,
     AppConfigService,
-    LogService,
-    IdentityUserModel,
-    IdentityUserService
+    LogService
 } from '@alfresco/adf-core';
 import { from, Observable, throwError } from 'rxjs';
 import { StartTaskCloudRequestModel } from '../models/start-task-cloud-request.model';
@@ -35,7 +32,6 @@ export class StartTaskCloudService {
     constructor(
         private apiService: AlfrescoApiService,
         private appConfigService: AppConfigService,
-        private identityUserService: IdentityUserService,
         private logService: LogService
     ) {}
 
@@ -60,98 +56,8 @@ export class StartTaskCloudService {
             );
     }
 
-    getUsers(): Observable<IdentityUserModel[]> {
-        const url = this.buildUserUrl();
-        const httpMethod = 'GET', pathParams = {}, queryParams = {}, bodyParam = {}, headerParams = {},
-            formParams = {}, authNames = [], contentTypes = ['application/json'], accepts = ['application/json'];
-
-        return from(this.apiService.getInstance().oauth2Auth.callCustomApi(
-                    url, httpMethod, pathParams, queryParams,
-                    headerParams, formParams, bodyParam, authNames,
-                    contentTypes, accepts, Object, null, null)
-                ).pipe(
-                    map((response: IdentityUserModel[]) => {
-                        return response;
-                    }),
-                catchError((err) => this.handleError(err))
-            );
-    }
-
-    getUserRoles(userId: string): Observable<RoleCloudModel[]> {
-        const url = this.buildRolesUrl(userId);
-        const httpMethod = 'GET', pathParams = {}, queryParams = {}, bodyParam = {}, headerParams = {},
-            formParams = {}, authNames = [], contentTypes = ['application/json'], accepts = ['application/json'];
-
-        return from(this.apiService.getInstance().oauth2Auth.callCustomApi(
-                    url, httpMethod, pathParams, queryParams,
-                    headerParams, formParams, bodyParam, authNames,
-                    contentTypes, accepts, Object, null, null)
-                ).pipe(
-                    map((response: RoleCloudModel[]) => {
-                        return response;
-                    }),
-                catchError((err) => this.handleError(err))
-            );
-    }
-
-    async getUsersByRolesWithCurrentUser(roleNames: string[]): Promise<IdentityUserModel[]> {
-        const filteredUsers: IdentityUserModel[] = [];
-        if (roleNames && roleNames.length > 0) {
-            const users = await this.getUsers().toPromise();
-
-            for (let i = 0; i < users.length; i++) {
-                const hasAnyRole = await this.userHasAnyRole(users[i].id, roleNames);
-                if (hasAnyRole) {
-                    filteredUsers.push(users[i]);
-                }
-            }
-        }
-
-        return filteredUsers;
-    }
-
-    async getUsersByRolesWithoutCurrentUser(roleNames: string[]): Promise<IdentityUserModel[]> {
-        const filteredUsers: IdentityUserModel[] = [];
-        if (roleNames && roleNames.length > 0) {
-            const currentUser = await this.identityUserService.getCurrentUserInfo().toPromise();
-            let users = await this.getUsers().toPromise();
-
-            users = users.filter((user) => { return user.username !== currentUser.username; });
-
-            for (let i = 0; i < users.length; i++) {
-                const hasAnyRole = await this.userHasAnyRole(users[i].id, roleNames);
-                if (hasAnyRole) {
-                    filteredUsers.push(users[i]);
-                }
-            }
-        }
-
-        return filteredUsers;
-    }
-
-    private async userHasAnyRole(userId: string, roleNames: string[]): Promise<boolean> {
-        const userRoles = await this.getUserRoles(userId).toPromise();
-        const hasAnyRole = roleNames.some((roleName) => {
-            const filteredRoles = userRoles.filter((userRole) => {
-                return userRole.name.toLocaleLowerCase() === roleName.toLocaleLowerCase();
-            });
-
-            return filteredRoles.length > 0;
-        });
-
-        return hasAnyRole;
-    }
-
     private buildCreateTaskUrl(appName: string): any {
         return `${this.appConfigService.get('bpmHost')}/${appName}-rb/v1/tasks`;
-    }
-
-    private buildUserUrl(): any {
-        return `${this.appConfigService.get('bpmHost')}/auth/admin/realms/springboot/users`;
-    }
-
-    private buildRolesUrl(userId: string): any {
-        return `${this.appConfigService.get('bpmHost')}/auth/admin/realms/springboot/users/${userId}/role-mappings/realm/composite`;
     }
 
     private buildRequestBody(taskDetails: any) {
